Use infer_schema_length param when reading clade assignments files
When testing against larger volumes of data, polars returned a schema
error when reading the clade assignments file generated by nextclade run.
By default, polars only looks at the first 100 rows to infer a file schema;
using a larger number of rows reduces the liklihood that it will infer
incorrectly.

The other alternative is to provide polars with a specific schema
or column type, but doing that increases the change of something
breaking due to an upstream (i.e., Nextrain) change.