Make config data_path optional for easier instantiation
The config object was originally designed to be used with the
script that assigns sequences to clades. We still have some
work to do to make the Config less specific to that original
usage, but to start with, let's make the data directory
optional and set it to a default value.