Add a sequence_metadata attribute to CladeTime
The first iteration of CladeTime contained a url_sequence_metadata
attribute that points to the S3 link for NextStrain's sequence
metadata file. This PR adds a sequence_metadata attribute that
supplies users with a Polars LazyFrame to the S3 file.

Note: doing the sorting and filtering on the S3 LazyFrame (i.e.,
without downloading the file first) saves time for those
interested in only a subset of the metadata (e.g., US only,
homo sapiens).